fix(gicv3): typecast operands to match data type
This corrects the MISRA violation C2012-10.3:
The value of an expression shall not be assigned to an object with a
narrower essential type or of a different essential type category.
The condition is explicitly checked against 0U, appending 'U' and
typecasting for unsigned comparison.

+++ drivers/arm/gic/v3/gicv3_main.c
@@ -1148,10 +1148,10 @@ void gicv3_raise_sgi(unsigned int sgi_num, gicv3_irq_group_t group,
 	assert((sgi_num >= MIN_SGI_ID) && (sgi_num < MIN_PPI_ID));
 
 	/* Extract affinity fields from target */
-	aff0 = MPIDR_AFFLVL0_VAL(target);
-	aff1 = MPIDR_AFFLVL1_VAL(target);
-	aff2 = MPIDR_AFFLVL2_VAL(target);
-	aff3 = MPIDR_AFFLVL3_VAL(target);
+	aff0 = (unsigned int)MPIDR_AFFLVL0_VAL(target);
+	aff1 = (unsigned int)MPIDR_AFFLVL1_VAL(target);
+	aff2 = (unsigned int)MPIDR_AFFLVL2_VAL(target);
+	aff3 = (unsigned int)MPIDR_AFFLVL3_VAL(target);
 
 	/*
 	 * Make target list from affinity 0, and ensure GICv3 SGI can target
